pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["GASC1211 <48170476+Gasc1211@users.noreply.github.com>"]
readme = "README.md"
include = ['models', 'utils', 'controllers']

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pyodbc = "^5.2.0"
azure-storage-queue = "^12.12.0"
python-dotenv = "^1.0.1"
firebase-admin = "^6.6.0"
pyjwt = "^2.10.0"
requests = "^2.32.3"
pytz = "^2024.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'server:dev'